+++ packages/vite/src/node/index.ts
@@ -16,7 +16,7 @@ export {
   minifySync,
   type MinifyOptions,
   type MinifyResult,
-} from 'rolldown/experimental'
+} from 'rolldown/utils'
 
 /** @deprecated - use `parse` instead */
 export const parseAst: typeof _parseAst = _parseAst